# Using vectorized loads (float2) for fp16 to improve performance