--- a/onnxruntime/contrib_ops/rocm/bert/fast_gelu_impl.cu
+++ b/onnxruntime/contrib_ops/rocm/bert/fast_gelu_impl.cu
@@ -83,7 +83,8 @@ __global__ void FastGeluKernel2(const half2 a, const half2 b, const half2 c, int
 }
 
 template <unsigned TPB>
-__global__ void FastGeluKernel4Bias(const half2 a, const half2 b, const half2 c, int input_length, int bias_length, const half* input, const half* bias, half* output) {
+__global__ void FastGeluKernel4Bias(const half2 a, const half2 b, const half2 c, int input_length, int bias_length, 
+		                    const half* input, const half* bias, half* output) {
   const int idx = blockIdx.x * TPB + threadIdx.x;
 
   const float2* input_cast = reinterpret_cast<const float2*>(input);
@@ -125,7 +126,8 @@ __global__ void FastGeluKernel4Bias(const half2 a, const half2 b, const half2 c,
 }
 
 template <unsigned TPB>
-__global__ void FastGeluKernel4(const half2 a, const half2 b, const half2 c, int input_length, const half* input, half* output) {
+__global__ void FastGeluKernel4(const half2 a, const half2 b, const half2 c, int input_length, 
+		                const half* input, half* output) {
   const int idx = blockIdx.x * TPB + threadIdx.x;
 
   const float2* input_cast = reinterpret_cast<const float2*>(input);
@@ -178,9 +180,11 @@ bool LaunchFastGeluKernel(const hipDeviceProp_t& prop, hipStream_t stream, int i
     const half2 B2 = __floats2half2_rn(B, B);
     const half2 C2 = __floats2half2_rn(C, C);
     if (bias == nullptr)
-      hipLaunchKernelGGL(HIP_KERNEL_NAME(FastGeluKernel4<blockSize>), dim3(gridSize), dim3(blockSize), 0, stream, A2, B2, C2, n, input, output);
+      hipLaunchKernelGGL(HIP_KERNEL_NAME(FastGeluKernel4<blockSize>), dim3(gridSize), dim3(blockSize), 0, 
+		                         stream, A2, B2, C2, n, input, output);
     else
-      hipLaunchKernelGGL(HIP_KERNEL_NAME(FastGeluKernel4Bias<blockSize>), dim3(gridSize), dim3(blockSize), 0, stream, A2, B2, C2, n, bias_length / 4, input, bias, output);
+      hipLaunchKernelGGL(HIP_KERNEL_NAME(FastGeluKernel4Bias<blockSize>), dim3(gridSize), dim3(blockSize), 0, 
+		                         stream, A2, B2, C2, n, bias_length / 4, input, bias, output);
   } else if (use_half2 && 0 == (bias_length & 1) && prop.major >= 7) {
     const int n = input_length / 2;
     const int gridSize = (n + blockSize - 1) / blockSize;