**Description**: Describe your changes.
Optimized LaunchFastGeluKernel for fp16 on AMD GPUs.

**Motivation and Context**
The performance gap is observed in FastGelu kernels with the [microbenchmark](https://github.com/microsoft/onnxruntime/blob/master/onnxruntime/python/tools/microbench/fast_gelu.py) on MI200 vs. A100. With the optimized kernels, we are able to see the significant performance improvement compared to A100.