@tianleiwu 2022-04-28
There are a few warnings from cpplint. Please format the code (see .clang-format in the root directory). One quick way is to rename fast_gelu_impl.cu to fast_gelu_impl.cc, then use Visual Studio to format the file, then undo the rename.