@hubertlu-tw 2022-05-04
The perf numbers in the following table were collected on MI200.
<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/hubertlu/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/hubertlu/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">

<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{mso-header-data:"&L&\0022Arial\0022&10&K008000 \[Public\]&1\#\000D";
	margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl65
	{color:red;}
.xl66
	{white-space:normal;}
-->

</head>

<body link="#0563C1" vlink="#954F72">



batch-size | seq-len | inter-dim | (Original) total time (us) over 110 runs | (Oprimized) total time (us) over 110 runs | Performance improvement (%) | Performance improvement (%)        after moving some computations to LaunchFastGeluKernel
-- | -- | -- | -- | -- | -- | --
1 | 128 | 3072 | 967 | 934 | 3.412616 | 7.238883144
1 | 128 | 4096 | 994 | 959 | 3.521127 | 3.822937626
1 | 384 | 3072 | 1134 | 1146 | -1.0582 | 2.116402116
1 | 384 | 4096 | 1320 | 1176 | 10.90909 | 10.83333333
1 | 512 | 3072 | 1307 | 1174 | 10.17598 | 10.4055088
1 | 512 | 4096 | 1559 | 1349 | 13.47017 | 13.34188582
4 | 128 | 3072 | 1315 | 1220 | 7.224335 | 10.87452471
4 | 128 | 4096 | 1557 | 1339 | 14.00128 | 14.25818882
4 | 384 | 3072 | 2762 | 2268 | 17.88559 | 18.42867487
4 | 384 | 4096 | 3487 | 2844 | 18.43992 | 18.81273301
4 | 512 | 3072 | 3478 | 2852 | 17.99885 | 18.94767108
4 | 512 | 4096 | 4476 | 3579 | 20.04021 | 20.57640751
8 | 128 | 3072 | 2032 | 1700 | 16.33858 | 16.38779528
8 | 128 | 4096 | 2519 | 2074 | 17.66574 | 18.26121477
8 | 384 | 3072 | 4952 | 3933 | 20.57754 | 20.88045234
8 | 384 | 4096 | 6403 | 5120 | 20.03748 | 20.16242386
8 | 512 | 3072 | 6427 | 5141 | 20.00934 | 20.63170997
8 | 512 | 4096 | 8382 | 6638 | 20.80649 | 21.22405154
32 | 128 | 3072 | 6391 | 5127 | 19.77781 | 20.24722266
32 | 128 | 4096 | 8388 | 6632 | 20.93467 | 21.28040057
32 | 384 | 3072 | 18114 | 14264 | 21.25428 | 21.78977586
32 | 384 | 4096 | 23999 | 18811 | 21.61757 | 22.08008667
32 | 512 | 3072 | 23983 | 18825 | 21.5069 | 21.9572197
32 | 512 | 4096 | 31790 | 24876 | 21.74898 | 22.21453287
64 | 128 | 3072 | 12265 | 9703 | 20.88871 | 21.52466368
64 | 128 | 4096 | 16172 | 12736 | 21.2466 | 21.71036359
64 | 384 | 3072 | 35649 | 28052 | 21.31056 | 21.69766333
64 | 384 | 4096 | 47332 | 37037 | 21.75061 | 22.22809093
64 | 512 | 3072 | 47322 | 37036 | 21.73619 | 22.17361904
64 | 512 | 4096 | 62939 | 49192 | 21.84178 | 22.37245587
128 | 128 | 3072 | 23958 | 18820 | 21.44586 | 21.90917439
128 | 128 | 4096 | 31775 | 24890 | 21.66798 | 22.15578285
128 | 384 | 3072 | 70674 | 55384 | 21.63455 | 21.90055749
128 | 384 | 4096 | 94078 | 73592 | 21.77555 | 22.28151109
128 | 512 | 3072 | 94023 | 73681 | 21.63513 | 22.18393372
128 | 512 | 4096 | 125250 | 97950 | 21.79641 | 22.11257485



</body>

</html>